Extracted DateRange TDG into separate class and made a proper enumerator method
